+++ Etyllica/src/br/com/etyllica/core/event/PointerEvent.java
@@ -16,9 +16,9 @@ public class PointerEvent {
 	
 	private PointerState state;
 	
-	private int x;
+	private float x;
 	
-	private int y;
+	private float y;
 	
 	private int amount;
 	
@@ -67,19 +67,19 @@ public class PointerEvent {
 		this.state = state;
 	}
 
-	public int getX() {
+	public float getX() {
 		return x;
 	}
 
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y = y;
 	}
