+++ Etyllica/src/br/com/etyllica/core/theme/ThemeFactory.java
@@ -4,6 +4,6 @@ import br.com.etyllica.gui.button.DefaultButton;
 
 public interface ThemeFactory {
 	
-	public DefaultButton createButton(int x, int y, int w, int h);
+	public DefaultButton createButton(float x, float y, float w, float h);
 
 }
