+++ Etyllica/src/br/com/etyllica/gui/ButtonList.java
@@ -78,7 +78,7 @@ public class ButtonList extends View{
 	}
 	
 	@Override
-	public void setOffset(int offsetX, int offsetY){
+	public void setOffset(float offsetX, float offsetY){
 		for(DefaultButton button: buttons){
 			button.setOffset(offsetX, offsetY);
 		}
